+++ CHANGES.md
@@ -2,9 +2,25 @@
 
 ### ? - ?
 
+##### Breaking Changes :mega:
+
+- Removed `tilesLoadingLowPriority`, `tilesLoadingMediumPriority`, and `tilesLoadingHighPriority` from `ViewUpdateResult`. Use `workerThreadTileLoadQueueLength` and `mainThreadTileLoadQueueLength` instead.
+
+##### Additions :tada:
+
+- Added `getOrientedBoundingBoxFromBoundingVolume` to the `Cesium3DTilesSelection` namespace.
+- Added `transform` and `toAxisAligned` methods to `OrientedBoundingBox`.
+- Switched to `libjpeg-turbo` instead of `stb` for faster jpeg decoding.
+- Added `getNumberOfTilesLoaded` method to `Tileset`.
+- Changed how `TilesetOptions::forbidHoles` works so that it loads much more quickly, while still guaranteeing there are no holes in the tileset.
+- Added `frameNumber` property to `ViewUpdateResult`.
+- Added getters for the `stride` and `data` fields of `AccessorView`.
+
 ##### Fixes :wrench:
 
-- Fixed a bug that prevented WebP and KTX2 textures from working in the common case where only the extension specified the `source` property, not the glTF's main Texture definition.
+- Fixed a bug that caused the `center` field of `AxisAlignedBox` to be incorrect.
+- Fixed a bug that caused the main thread to sometimes load low-priority tiles before high-priority ones. This could result in much longer waits than necessary for a tileset's appropriate level-of-detail to be shown.
+- Fixed a bug that prevented WebP and KTX2 textures from working in the common case where only the extension specified the `source` property, not the glTF's main `Texture` definition.
 
 ### v0.22.1 - 2023-03-06